Seed should never be int

+++ src/main/java/cn/nukkit/Server.java
@@ -373,7 +373,7 @@ public class Server {
 
         for (String name : ((Map<String, Object>) this.getConfig("worlds", new HashMap<>())).keySet()) {
             if (!this.loadLevel(name)) {
-                int seed = (int) this.getConfig("worlds." + name + ".seed", Long.valueOf(System.currentTimeMillis()).intValue());
+                long seed = (long) this.getConfig("worlds." + name + ".seed", System.currentTimeMillis());
 
                 Map<String, Object> options = new HashMap<>();
                 String[] opts = ((String) this.getConfig("worlds." + name + ".generator", Generator.getGenerator("default").getSimpleName())).split(":");
